+++ internal/pkg/instrumentation/manager.go
@@ -13,7 +13,6 @@ import (
 	"github.com/cilium/ebpf/link"
 	"github.com/cilium/ebpf/rlimit"
 
-	"go.opentelemetry.io/collector/pdata/ptrace"
 	"go.opentelemetry.io/otel/trace"
 
 	dbSql "go.opentelemetry.io/auto/internal/pkg/instrumentation/bpf/database/sql"
@@ -59,7 +58,6 @@ type Manager struct {
 	exe             *link.Executable
 	td              *process.TargetDetails
 	runningProbesWG sync.WaitGroup
-	telemetryCh     chan ptrace.ScopeSpans
 	currentConfig   Config
 	probeMu         sync.Mutex
 	state           managerState
@@ -68,13 +66,12 @@ type Manager struct {
 // NewManager returns a new [Manager].
 func NewManager(logger *slog.Logger, otelController *opentelemetry.Controller, globalImpl bool, cp ConfigProvider, version string) (*Manager, error) {
 	m := &Manager{
-		logger:         logger,
-		version:        version,
-		probes:         make(map[probe.ID]probe.Probe),
-		otelController: otelController,
-		globalImpl:     globalImpl,
-		cp:             cp,
-		telemetryCh:    make(chan ptrace.ScopeSpans),
+		logger:          logger,
+		version:         version,
+		probes:          make(map[probe.ID]probe.Probe),
+		otelController:  otelController,
+		globalImpl:      globalImpl,
+		cp:              cp,
 	}
 
 	err := m.registerProbes()
@@ -227,7 +224,7 @@ func (m *Manager) runProbe(p probe.Probe) {
 	m.runningProbesWG.Add(1)
 	go func(ap probe.Probe) {
 		defer m.runningProbesWG.Done()
-		ap.Run(m.telemetryCh)
+		ap.Run(m.otelController.Trace)
 	}(p)
 }
 
@@ -302,9 +299,6 @@ func (m *Manager) Run(ctx context.Context) error {
 		done <- errors.Join(err, ctx.Err())
 	}()
 
-	for e := range m.telemetryCh {
-		m.otelController.Trace(e)
-	}
 	return <-done
 }
 
@@ -322,7 +316,6 @@ func (m *Manager) Stop() error {
 
 	// Wait for all probes to stop before closing the chan they send on.
 	m.runningProbesWG.Wait()
-	close(m.telemetryCh)
 
 	m.state = managerStateStopped
 	return err